refactored_auth_file_slices
