+++ Clients/src/infrastructure/api/customAxios.ts
@@ -23,7 +23,7 @@
 import axios, { AxiosError, InternalAxiosRequestConfig } from "axios";
 import { store } from "../../application/redux/store";
 import { ENV_VARs } from "../../../env.vars";
-import { clearAuthState, setAuthToken } from "../../application/authentication/authSlice";
+import { clearAuthState, setAuthToken } from "../../application/redux/auth/authSlice";
 import { AlertProps } from "../../domain/interfaces/iAlert";
 
 const performLogout = () => {
