+++ Clients/src/presentation/components/ProtectedRoute/index.tsx
@@ -2,7 +2,7 @@ import { useSelector } from "react-redux";
 import { Navigate, useLocation } from "react-router-dom";
 import { ComponentType, useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
-import { setUserExists } from "../../../application/authentication/authSlice";
+import { setUserExists } from "../../../application/redux/auth/authSlice";
 import { getAllEntities } from "../../../application/repository/entity.repository"; // Import the checkUserExists function
 import CustomizableToast from "../../vw-v2-components/Toast";
