+++ Clients/src/presentation/pages/Authentication/Login/index.tsx
@@ -8,8 +8,8 @@ import { useNavigate } from "react-router-dom";
 import { loginUser } from "../../../../application/repository/entity.repository";
 import { logEngine } from "../../../../application/tools/log.engine";
 import { useDispatch } from "react-redux";
-import { setAuthToken } from "../../../../application/authentication/authSlice";
-import { setExpiration } from "../../../../application/authentication/authSlice";
+import { setAuthToken } from "../../../../application/redux/auth/authSlice";
+import { setExpiration } from "../../../../application/redux/auth/authSlice";
 import CustomizableToast from "../../../vw-v2-components/Toast";
 import Alert from "../../../components/Alert";
 import { ENV_VARs } from "../../../../../env.vars";
