+++ Clients/src/presentation/pages/Authentication/RegisterAdmin/index.tsx
@@ -24,7 +24,7 @@ import { useDispatch } from "react-redux";
 import {
   setUserExists,
   setAuthToken,
-} from "../../../../application/authentication/authSlice";
+} from "../../../../application/redux/auth/authSlice";
 import { VerifyWiseContext } from "../../../../application/contexts/VerifyWise.context";
 
 // Initial state for form values
